# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallet API",
        "description": "API documentation for Wallet transactions",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/scoreboard": {
            "post": {
                "tags": [
                    "ScoreBoard"
                ],
                "summary": "Save or update scoreboard",
                "operationId": "ac5116d41593d2f7ddc58dcd78229c48",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "google_user_id",
                                    "sub_category_id",
                                    "total_videos",
                                    "quiz_practice",
                                    "test_rank"
                                ],
                                "properties": {
                                    "google_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "sub_category_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "total_videos": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "quiz_practice": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "test_rank": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Scoreboard saved successfully"
                    },
                    "400": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/api/scoreboard/{userId}": {
            "get": {
                "tags": [
                    "ScoreBoard"
                ],
                "summary": "Get scoreboard by user ID",
                "operationId": "dce6cc83de98c5a4900583d188b41605",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Scoreboard retrieved successfully"
                    },
                    "404": {
                        "description": "No scoreboard found for this user"
                    }
                }
            }
        },
...
        "/api/wallet/add": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Add coins to the wallet",
                "operationId": "4d5a141fc4de4c0de156dde6914449d7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "google_user_id",
                                    "coin",
                                    "method",
                                    "date",
                                    "transaction_id",
                                    "amount"
                                ],
                                "properties": {
                                    "google_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "coin": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "PayPal"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-31"
                                    },
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "txn_123456"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet credited successfully"
                    },
                    "400": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/wallet/charges": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Deduct coins from the wallet",
                "operationId": "3e81695396c7e6c53e279b8f37458e18",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "google_user_id",
                                    "coin",
                                    "method",
                                    "date",
                                    "transaction_id",
                                    "amount"
                                ],
                                "properties": {
                                    "google_user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "coin": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "Bank Transfer"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-31"
                                    },
                                    "transaction_id": {
                                        "type": "string",
                                        "example": "txn_654321"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 5.25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Wallet debited successfully"
                    },
                    "400": {
                        "description": "Insufficient coins or validation errors"
                    }
                }
            }
        }
    },
...
}